Add Docker support and restructure project for scalability
This commit introduces Docker support with a Dockerfile, docker-compose file, and an image push script. It also restructures project files, moving components to a dedicated "pages" folder and improving routing with `react-router-dom`. Minor updates include adjustments to `vite.config.js`, `.env.sample`, and package dependencies for better development and deployment practices.
