+++ .env.sample
@@ -1,2 +1,8 @@
 VITE_CHESSHUB_CORE_BASEURL=""
 VITE_CHESS_SCORESHEET_DIGITALIZATION_BASEURL=""
+
+# Docker
+IMAGES=("chesshub-ui")
+VERSION=1.0.0
+DOCKER_USERNAME=bkostka
+DOCKER_PASSWORD=
