+++ etc/nginx/nginx.conf
@@ -1,0 +1,13 @@
+server {
+    listen 80;
+
+    index index.html;
+
+    location / {
+        try_files $uri /index.html;  # Unbekannte Routen gehen auf index.html zurück
+    }
+
+    # Optional: gzip aktivieren, um die Leistung zu verbessern
+    gzip on;
+    gzip_types text/plain text/css application/json application/javascript text/xml application/xml application/xml+rss text/javascript;
+}
