+++ package.json
@@ -17,6 +17,7 @@
     "react": "^19.0.0",
     "react-dom": "^19.0.0",
     "react-router": "^7.5.0",
+    "react-router-dom": "^7.6.0",
     "tanstack": "^1.0.0"
   },
   "devDependencies": {
