+++ src/App.jsx
@@ -1,8 +1,9 @@
-import '../styles/App.css'
-import { Registration } from "./Registration";
+import './styles/App.css'
+import { Registration } from "./pages/Registration.jsx";
 import {BrowserRouter, Route, Routes} from "react-router";
-import { Login } from "./Login";
-import {UploadImage} from "./UploadImage.jsx";
+import { Login } from "./pages/Login.jsx";
+import {UploadImage} from "./pages/UploadImage.jsx";
+import {Home} from "./pages/Home.jsx";
 
 export function App() {
 
@@ -11,6 +12,7 @@ export function App() {
           <BrowserRouter>
               <main>
                   <Routes>
+                      <Route path="/" element={< Home />} />
                       <Route path="/auth/register" element={< Registration />} />
                       <Route path="/auth/signup" element={< Registration />} />
                       <Route path="/auth/login" element={< Login />} />
