+++ src/index.jsx
@@ -1,7 +1,7 @@
 import { StrictMode } from 'react'
 import { createRoot } from 'react-dom/client'
 import './styles/index.css'
-import { App } from './components/App.jsx'
+import { App } from './App.jsx'
 
 createRoot(document.getElementById('root')).render(
   <StrictMode>
